build: adopt the harness security toolchain
Add gosec, govulncheck, and staticcheck as go.mod tool directives and wire
fmt-check, vet, staticcheck, gosec, and govulncheck into make lint / vuln /
secure, matching harness. There is no vendor tree here, so the vendor targets
are omitted.

make vet checks BOTH platforms. Roughly half this module sits behind
//go:build linux, so a host-only vet silently skips the namespace, Landlock,
seccomp, nftables, and cgroup code — the code most worth checking. That linux
pass immediately earned its keep by catching test helpers that darwin's
staticcheck had reported as dead.

Removes genuinely dead code the new linters found: underWritableRoot,
hasDenyEntry, assertGuarantees, Executor.revoke, FSRule.writable, and
enumerateGrantPathHandle, none of which had any caller.

Fixes one real defect in the file-identity conversion: Stat_t.Dev is int32 on
darwin and uint64 on linux, so the fixed-width cast was wrong on one of them.
Formatting each field in its native width is correct on both.

The remaining five gosec findings are annotated rather than silenced blindly —
launching a command is this module's purpose, 0700 is correct for a directory,
and the planted test sentinel is the opposite of a credential.

Adding the tools raised golang.org/x/sys from v0.40.0 to v0.45.0, since minimal
version selection takes the maximum across the whole module graph. Both
platforms build and the full race suite passes on the newer version.

+++ internal/policy/identity_unix.go
@@ -21,5 +21,9 @@ func fileInfoIdentity(info os.FileInfo) (string, error) {
 	if !ok {
 		return "", fmt.Errorf("sandbox: filesystem identity unavailable")
 	}
-	return fmt.Sprintf("%d:%d:%d", uint64(stat.Dev), uint64(stat.Ino), uint32(info.Mode().Type())), nil
+	// No fixed-width conversion: Stat_t.Dev is int32 on darwin and uint64 on
+	// linux, so any cast would be wrong on one of them. %d formats each field in
+	// its native width, and the result is an opaque identity only ever compared
+	// for equality against another identity captured on the same host.
+	return fmt.Sprintf("%d:%d:%d", stat.Dev, stat.Ino, info.Mode().Type()), nil
 }
